update models with new fragment for timeslot
