+++ Core/GraphQL/API/sell_getFulfilmentPoints.graphql.swift
@@ -27,6 +27,7 @@ public extension ApolloType {
       var document: String = operationDefinition
       document.append("\n" + FragmentFulfilmentPoint.fragmentDefinition)
       document.append("\n" + FragmentFulfilmentPointCategory.fragmentDefinition)
+      document.append("\n" + FragmentTimeslot.fragmentDefinition)
       return document
     }
