+++ Core/GraphQL/API/sell_getProducts.graphql.swift
@@ -26,6 +26,7 @@ public extension ApolloType {
       document.append("\n" + FragmentProduct.fragmentDefinition)
       document.append("\n" + FragmentFulfilmentPoint.fragmentDefinition)
       document.append("\n" + FragmentFulfilmentPointCategory.fragmentDefinition)
+      document.append("\n" + FragmentTimeslot.fragmentDefinition)
       return document
     }
